requirements.txt
# Core Web Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==3.0.1

# Deep Learning & PyTorch
torch==2.1.2
torchvision==0.16.2
timm==0.9.16

# Data Processing
numpy==1.26.4
pandas==2.1.1
scipy==1.11.4
scikit-learn==1.3.2
scikit-image==0.21.0
joblib==1.3.2
tqdm==4.66.1

# Computer Vision
opencv-python==4.8.1.78
albumentations==1.3.1
Pillow==10.1.0

# Explainability
grad-cam==1.4.6

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
